requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.2.0
python-multipart>=0.0.9
apscheduler>=3.10.4
httpx>=0.27.0
python-dotenv>=1.0.0
croniter>=2.0.1
email-validator>=2.1.0
